docs(asset-events-panel): rewrite docs in new format
